refactor: 체크리스트 답변 관리 로직을 useChecklistAnswer 로 리팩토링
